Move data op resolving to the indexing layer
