+++ src/server/pkg/storage/fileset/index/util.go
@@ -1,6 +1,10 @@
 package index
 
-import "sort"
+import (
+	"sort"
+
+	"github.com/pachyderm/pachyderm/src/server/pkg/storage/chunk"
+)
 
 // Perm calls f with each permutation of a.
 func Perm(a []rune, f func([]rune)) {
@@ -29,4 +33,43 @@ func Generate(s string) []string {
 	})
 	sort.Strings(fileNames)
 	return fileNames
+}
+
+func resolveDataOps(idx *Index) {
+	if idx.FileOp.DataRefs == nil {
+		return
+	}
+	dataRefs := idx.FileOp.DataRefs
+	offset := dataRefs[0].OffsetBytes
+	size := dataRefs[0].SizeBytes
+	for _, dataOp := range idx.FileOp.DataOps {
+		bytesLeft := dataOp.SizeBytes
+		for size <= bytesLeft {
+			dataOp.DataRefs = append(dataOp.DataRefs, newDataRef(dataRefs[0].ChunkInfo, offset, size))
+			bytesLeft -= size
+			dataRefs = dataRefs[1:]
+			if len(dataRefs) == 0 {
+				return
+			}
+			offset = dataRefs[0].OffsetBytes
+			size = dataRefs[0].SizeBytes
+		}
+		dataOp.DataRefs = append(dataOp.DataRefs, newDataRef(dataRefs[0].ChunkInfo, offset, bytesLeft))
+		offset += bytesLeft
+		size -= bytesLeft
+	}
+}
+
+func newDataRef(chunkInfo *chunk.ChunkInfo, offset, size int64) *chunk.DataRef {
+	return &chunk.DataRef{
+		ChunkInfo:   chunkInfo,
+		OffsetBytes: offset,
+		SizeBytes:   size,
+	}
+}
+
+func unresolveDataOps(idx *Index) {
+	for _, dataOp := range idx.FileOp.DataOps {
+		dataOp.DataRefs = nil
+	}
 }
