+++ src/server/pkg/storage/fileset/util.go
@@ -203,42 +203,3 @@ func getDataRefs(dataOps []*index.DataOp) []*chunk.DataRef {
 	}
 	return dataRefs
 }
-
-func resolveDataOps(idx *index.Index) {
-	if idx.FileOp.DataRefs == nil {
-		return
-	}
-	dataRefs := idx.FileOp.DataRefs
-	offset := dataRefs[0].OffsetBytes
-	size := dataRefs[0].SizeBytes
-	for _, dataOp := range idx.FileOp.DataOps {
-		bytesLeft := dataOp.SizeBytes
-		for size <= bytesLeft {
-			dataOp.DataRefs = append(dataOp.DataRefs, newDataRef(dataRefs[0].ChunkInfo, offset, size))
-			bytesLeft -= size
-			dataRefs = dataRefs[1:]
-			if len(dataRefs) == 0 {
-				return
-			}
-			offset = dataRefs[0].OffsetBytes
-			size = dataRefs[0].SizeBytes
-		}
-		dataOp.DataRefs = append(dataOp.DataRefs, newDataRef(dataRefs[0].ChunkInfo, offset, bytesLeft))
-		offset += bytesLeft
-		size -= bytesLeft
-	}
-}
-
-func unresolveDataOps(idx *index.Index) {
-	for _, dataOp := range idx.FileOp.DataOps {
-		dataOp.DataRefs = nil
-	}
-}
-
-func newDataRef(chunkInfo *chunk.ChunkInfo, offset, size int64) *chunk.DataRef {
-	return &chunk.DataRef{
-		ChunkInfo:   chunkInfo,
-		OffsetBytes: offset,
-		SizeBytes:   size,
-	}
-}
